+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # JCM v2.2.0-beta.3 for MTR 4.0.4 has been released!
 
 ## Additions
-- Add `RenderManager.drawModel(model: Model, matrices: Matrices?)` as an alternative way to draw model with RenderManager.
+- Add `RenderManager.drawModel(model: Model | DynamicModelHolder, matrices: Matrices?)` as an alternative way to draw model with RenderManager.
+- Add `VehicleScriptContext.drawCarModel` / `EyecandyScriptContext.drawModel` with `DynamicModelHolder` overload instead of `Model`.
 - Add more overload for `GraphicsTexture.upload()`, allowing for uploading selected portion of image to conserve performance.
 
 ## Changes